# resources/api-spec-v4.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
  - description: Postman mock server
    url: https://a00459c1-e91f-47ea-9c25-bd9e5268f373.mock.pstmn.io
info:
  description: API specification for the SENG302 2021 project
  version: 4.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/businesses/{businessId}/products/{productId}/images':
    post:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Upload a new image for some photo
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business to search catalogue of
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to try to find in catalogue
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                filename:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Image created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageId:
                    type: integer
                    description: The ID of the image that has just been created
                    example: 1234
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to add an image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{businessId}/products/{productId}/images/{imageId}':
    get:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Get the image with the given image ID, product, and business.
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to update primary image for
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          description: ID of the image to set as being the primary image for product
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Primary image successfully updated
          content:
            image/png or image/jpeg or image/gif:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    delete:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Delete an image of a product
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to delete image for
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          description: ID of the image to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to delete an image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{businessId}/products/{productId}/images/{imageId}/thumbnail':
    get:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Get the image thumbnail with the given image ID, product, and business.
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to update primary image for
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          description: ID of the image to set as being the primary image for product
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Primary image successfully updated
          content:
            image/png or image/jpeg or image/gif:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{businessId}/products/{productId}/images/{imageId}/makeprimary':
    put:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Set an existing image to be the new primary image for some product
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to update primary image for
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          description: ID of the image to set as being the primary image for product
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Primary image successfully updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to update the primary image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  '/listings/{listingId}':
    get:
      tags:
        - 'U30: Individual Full Sale Listing'
        - 'Listings / sales'
      summary: 'Retrieve a specific sale listing by the sale listing ID'
      parameters:
        - name: listingId
          in: path
          description: ID of the specific sale listing to retrieve
          required: true
          schema:
            type: integer
            example: 17
      security:
        - CookieAuth:
            - user
      responses:
        '200':
          description: Successfully retrieved sale listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  '/listings/search':
    get:
      tags:
        - 'U29: Browse/Search Sale listings'
        - 'Listings / sales'
      summary: 'Search for sale listings based on ...'
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: searchQuery
          in: query
          description: 'Criteria to search sale listings for, can be product name, business name, or location (city/country)'
          required: false
          schema:
            type: string
        - name: businessType
          in: query
          description: 'Criteria to filter sale listings for, by business type, price range, closing date range'
          required: false
          schema:
            type: string
            enum:
              - 'Accommodation and Food Services'
              - 'Retail Trade'
              - 'Charitable organisation'
              - 'Non-profit organisation'
        - name: maxPrice
          in: query
          description: 'Criteria to filter sale listings for, by max price'
          required: false
          schema:
            type: integer
        - name: minPrice
          in: query
          description: 'Criteria to filter sale listings for, by min price'
          required: false
          schema:
            type: integer
        - name: earliestClosingDate
          in: query
          description: 'Criteria to filter sale listings for, by earliest closing date'
          required: false
          schema:
            type: string
        - name: latestClosingDate
          in: query
          description: 'Criteria to filter sale listings for, by latest closing date'
          required: false
          schema:
            type: string
        - name: pageNum
          in: query
          description: 'The page number to be retrieved for frontend'
          required: true
          schema:
            type: integer
        - name: perPage
          in: query
          description: 'The quantity of the items to be displayed per page'
          required: true
          schema:
            type: integer
        - name: sortBy
          in: query
          description: 'The field name chose to sort the search result. (productName, sellerName, location, price, closingDate)'
          required: true
          schema:
            type: string
        - name: orderBy
          in: query
          description: 'The ordering direction(asc, desc)'
          required: true
          schema:
            type: string
      responses:
          '200':
            description: 'Request successful, a list of sale listings is returned'
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Listing'
          '401':
            $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...
  schemas:
...
    Listing:
      type: object
      description: A listing for the sale of some product from a business's inventory
      required:
        - inventoryItemId
        - quantity
        - price
      properties:
        id:
          type: integer
          example: 57
          readOnly: true
          description: A unique (within the business) identifer for this listing - autogenerated by the system
        sold:
          type: boolean
          example: true
          description: Indicator whether the sale listing has been purchased from the business
        inventoryItem:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/InventoryItem'
        inventoryItemId:
          type: integer
          example: 101
          writeOnly: true
        quantity:
          type: integer
          example: 3
          description: Must be greater than zero, and less than or equal to the quantity of the correpsonding inventory entry
        price:
          type: number
          format: double
          description: Price of the listing
          example: 17.99
        moreInfo:
          type: string
          example: 'Seller may be willing to consider near offers'
          description: An optional field for further information about the listing
        created:
          type: string
          format: date-time
          example: '2021-07-14T11:44:00Z'
          readOnly: true
          description: Date and time when the listing was created
        closes:
          type: string
          format: date-time
          example: '2021-07-21T23:59:00Z'
          description: Date and time when the listing closes (defaults to expiry date of related inventory item)
        purchased:
          type: string
          format: date-time
          example: '2021-07-14T11:44:00Z'
          readOnly: true
          description: Date and time when the listing was purchased
        purchaserId:
          type: integer
          example: 1
          description: The Id of the user who purchased the sale listing
        country:
          type: string
          example: 'New Zealand'
          description: The country where the business that owns the sale listing exists.
        businessId:
          type: integer
          example: 1
          description: The businessId of the business that owns the sale listing.
        purchasedBy:
          readOnly: true
          allOf:
              - $ref: '#/components/schemas/User'
          description: The user who purchased this sale listing.
        likedByUsers:
          type: array
          readOnly: true
          description: '(Implemented in U30): A list of users who have liked the listing'
          items:
            $ref: '#/components/schemas/User'
    InventoryItem:
      type: object
      description: Entity representing an instance of a stock of some product
      required:
        - productId
        - quantity
        - expires
      properties:
        id:
          type: integer
          example: 101
          readOnly: true
          description: A unique (within the inventory) identifer for this item - autogenerated by the system
        product:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/Product'
        productId:
          writeOnly: true
          type: string
          example: "WATT-420-BEANS"
          description: Product Code of corresponding product in business's catalogue. Notice that the ID is given when writing, but the full object is returned when reading
        quantity:
          type: integer
          example: 4
          description: How many items of stock are present
        pricePerItem:
          type: number
          format: double
          description: Price to purchase a single item of the product
          example: 6.50
        totalPrice:
          type: number
          format: double
          description: Price for purchasing the entire stock of the product
          example: 21.99
        manufactured:
          type: string
          format: date
          description: The date on which the product was prepared / manufactured
        sellBy:
          type: string
          format: date
          description: The date by which the product ought to be sold
        bestBefore:
          type: string
          format: date
          description: 'The "best before" date of the product'
        expires:
          type: string
          format: date
          description: The expiry date of the product
    Image:
      type: object
      description: Entity representing an image stored on the server
      required:
        - filename
        - thumbnailFilename
      properties:
        id:
          type: integer
          example: 1234
          readOnly: true
          description: A unique identifier (system-wide) to represent this image
        filename:
          type: string
          example: "/media/images/23987192387509-123908794328.png"
          description: Relative URI to full size image. A fully qualified URI (e.g www.wasteless.com/media/images/...) should NOT be given.
        thumbnailFilename:
          type: string
          example: "/media/images/23987192387509-123908794328_thumbnail.png"
          description: Relative URI to thumbnail of image. A fully qualified URI (e.g www.wasteless.com/media/images/...) should NOT be given.
    Product:
      type: object
      description: Entity for catalogue items in a product catalogue
      required:
        - name
      properties:
        id:
          type: string
          example: 'WATT-420-BEANS'
          description: >
            Developers can choose whether or not to make this a required field (i.e user
            specified vs. auto-generated). In the case that the team automatically
            generates the IDs, they should still allow this field to be present when
            creating (POST) or modifying (PUT) a product, and then simply ignore its
            value (i.e do not return a 400 code if it is present but not needed). An
            ID only needs to be unique within its catalogue, not system-wide.
        name:
          type: string
          description: The full name of the product
          example: "Watties Baked Beans - 420g can"
        description:
          type: string
          description: A desription of the product
          example: "Baked Beans as they should be."
        manufacturer:
          type: string
          description: The manufacturer of the product
          example: "Heinz Wattie's Limited"
        recommendedRetailPrice:
          type: number
          format: double
          example: 2.20
          description: The recommended retail price of the product, given in whatever currency is in use at the business's location.
        created:
          type: string
          format: date-time
          readOnly: true
          description: The date and time when this product was created. This field is generated automatically by the server.
        images:
          type: array
          readOnly: true
          description: List of images of a product (implemented in U16).
          items:
            $ref: '#/components/schemas/Image'
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: "The ID of the 'primary' administrator of the business. This is automatically given to the creator of the business."
        name:
          type: string
          example: 'Lumbridge General Store'
        description:
          type: string
          example: 'A one-stop shop for all your adventuring needs'
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          type: string
          enum:
            - Accommodation and Food Services
            - Retail Trade
            - Charitable organisation
            - Non-profit organisation
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
    Address:
      type: object
      description: Entity representing some physical address
      required:
        - country
      properties:
        streetNumber:
          type: string
          example: '3/24'
        streetName:
          type: string
          example: 'Ilam Road'
        suburb:
          type: string
          example: 'Upper Riccarton'
        city:
          type: string
          example: 'Christchurch'
        region:
          type: string
          example: 'Canterbury'
        country:
          type: string
          example: 'New Zealand'
        postcode:
          type: string
          example: '90210'
    User:
      type: object
      description: Basic user entity
      required:
        - firstName
        - lastName
        - email
        - dateOfBirth
        - homeAddress
        - password
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        middleName:
          type: string
          example: Hector
        nickname:
          type: string
          example: Jonny
        bio:
          type: string
          example: Likes long walks on the beach
        email:
          type: string
          format: email
          example: johnsmith99@gmail.com
        dateOfBirth:
          type: string
          format: date
          example: '1999-04-27'
        phoneNumber:
          type: string
          example: '+64 3 555 0129'
        homeAddress:
          $ref: '#/components/schemas/Address'
        password:
          type: string
          example: '1337-H%nt3r2'
          format: password
          writeOnly: true
          description: Should not be stored in plain text (i.e hash it)
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date and time when the account was created
        role:
          type: string
          readOnly: true
          description: '(Implemented in U4): This property should only be shown to Global Application Admins'
          enum:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
        businessesAdministered:
          type: array
          readOnly: true
          description: '(Implemented in U6): A list of businesses for which the user is an administrator'
          items:
            $ref: '#/components/schemas/Business'
            
        likedSaleListings:
          type: array
          readOnly: true
          description: '(Implemented in U30): A list of sale listings the user has liked'
          items:
            $ref: '#/components/schemas/Listing'
...